% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-spread-graph.R
\name{center_spread_graph}
\alias{center_spread_graph}
\title{Main graphic for center/spread app}
\usage{
center_spread_graph(
  formula,
  data,
  yrange = NULL,
  labels = NULL,
  annots = list(prob_level = 0.95, show_median = TRUE, show_summary_interval = TRUE,
    show_mean = TRUE, show_sd = TRUE, show_ci = TRUE, show_violin = TRUE)
)
}
\arguments{
\item{formula}{specifying the response and explanatory variables. It is assumed that the
response variable is quantitative. If it's a 0/1 variable, the labels can be specified.}

\item{data}{the data frame containing the variables in the formula.}

\item{yrange}{the vertical range of the graphic (optional)}

\item{labels}{For a 0/1 response variable, categorical labels (if any) to be used for the 0 and 1}

\item{annots}{a list with elements
values.}
}
\description{
Makes the standard graphic for the Center & Spread app. The response variable
is expected to be quantitative, the explanatory variable categorical.
}
