% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_select_data.R
\name{data_tab}
\alias{data_tab}
\title{Functions for constructing the data UI}
\usage{
data_tab(package_list = default_packages_list, covar = TRUE,
  covar2 = FALSE)
}
\arguments{
\item{package_list}{A named list whose values are character strings containing
the packages to be included and whose names are  the user-facing strings for the
select widget. See `default_packages_list` for an example.}

\item{covar}{Flag: whether to add a selector for a covariate}

\item{covar2}{Flag: whether to add a selector for a second covariate (if `covar=TRUE`)}
}
\description{
Creates the tab with controls for selecting package, frame, response,
and  explanatory variables.
}
