% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicotomize.R
\name{dichotomize}
\alias{dichotomize}
\title{dichotomize a variable}
\usage{
dichotomize(x, pop = x, force = TRUE, to_numeric = FALSE)
}
\arguments{
\item{x}{the variable (as a vector) to be split}

\item{pop}{an (optional) reference set of data (typically from a much larger sample)
that determines the dichotomization level.}

\item{force}{a logical flag. If `TRUE`, even
numerical variables will be split.}

\item{to_numeric}{a logical flag. If `TRUE`, dichotomized factors
will be reported as 0/1 numbers with the levels in the attributes.}
}
\description{
Split a variable into two levels. Logical variables
will be converted to a TRUE/FALSE factor. Numerical variables
will be split at the median and returned as a factor. Factors will be
lumped into the most populated and all the others.
}
\details{
When `to_numeric` is true, the names of the dichotomized levels
will be reported as an attribute `levels` in the result. This can be used to label the
plot. (It will be `NULL` if the variable was already numeric.)
}
