% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_calculations.R
\name{F_main_calc}
\alias{F_main_calc}
\title{Main calculations for the F app}
\usage{
F_main_calc(
  formula,
  data,
  yrange = NULL,
  labels = NULL,
  show_model = TRUE,
  sd = TRUE
)
}
\arguments{
\item{formula}{of form A ~ B or A ~ B + C. You can also
use,  say, ns(B,2) or * instead of +. But at most two
explanatory variables.}

\item{data}{a data frame with the sample of data to display}

\item{yrange}{a y range to override that implied by the data}

\item{labels}{Optionally, a length-2 character giving the labels for 0 and 1
when the response variable is an indicator variable.}

\item{show_model}{flag to draw the model}

\item{sd}{logical flag to show the plus-or-minus 1 sd in  the side plot}

\item{R2}{logical flag to show the R-squared in the side plot}

\item{F}{logical flag to show the F value in the side plot}
}
\value{
A list with the two plots (P1 and P2),
the fitted model, and the stats.
}
\description{
Main calculations for any app consist of ...
\itemize{
\item (1) The main plot \code{main}
\item (2) The  side plot \code{side}. Can be null is none is needed.
\item (3) \code{stats} A data  frame or list with statistics on the system
list(main = P1 \%>\% gf_theme(legend.position  =  "left"),
}
}
\details{
A basic plot showing two or three variables:
\itemize{
\item a response
\item an explanatory variable
\item a covariate (optional)
Constructs two graphics, one showing the data
and (optionally) the fitted model, the other showing
showing the side plot comparing the distribution
of raw to  model values
}
}
\examples{
tmp <- F_app_plot(wage ~ splines::ns(educ, 2)*sex,  data = mosaicData::CPS85)
}
