% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_side_plot.R
\name{F_side_plot}
\alias{F_side_plot}
\title{Graphics for the 2020 revisions of Little Apps}
\usage{
F_side_plot(raw, fitted, explan = "bogus", dflex = 1, violin = FALSE,
  R2 = TRUE, F = TRUE, sd = TRUE)
}
\arguments{
\item{raw}{the response variable as a vector}

\item{fitted}{the fitted model values as a vector}

\item{explan}{the explanatory variable as a vector}

\item{violin}{if `TRUE` show a violin plot.}

\item{F}{if `TRUE`  show n and F statistic}

\item{sd}{if `TRUE`, show the interval of plus-or-minus one
standard deviation}

\item{R}{if `TRUE` show R  and R-squared in the graph.}
}
\description{
plot_values() creates the auxilliary graphic showing
the distribution of response and model values. Note
that covariates are not displayed even if they  have been
included in the model values. This simplifies the display.
}
\examples{
with(mosaicData::Galton,
side_plot_R(height, fitted(lm(height ~ father)),
  explan = father))
}
