% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_panels.R
\name{graph_panel}
\alias{graph_panel}
\title{panels for the UI}
\usage{
graph_panel()
}
\description{
Functions to make tabs for the UI: `graph_panel()`,  `compare_panel()`, `stats_panel()`,
`debug_panel()`
}
