% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_main.R
\name{miniTabstripPanel2}
\alias{miniTabstripPanel2}
\title{Layout for the main app}
\usage{
miniTabstripPanel2(..., id = NULL, selected = NULL, between = NULL)
}
\arguments{
\item{...}{tabs to  be inserted into  the app}

\item{top_controls}{contents of the top block of controls}
}
\description{
A Little App  consists of a top block of controls that stay visible aways, and
a set of tabs that implement different views: e.g. select data, compare current
to frozen plot, show statistics, show a description of the app, ...
}
