% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to_closest.R
\name{round_to_closest}
\alias{round_to_closest}
\title{"Round" a vector of numerical values to the closest element of a
set of discrete  values.}
\usage{
round_to_closest(x, discrete)
}
\arguments{
\item{x}{the vector of values}

\item{discrete}{a discrete set of numerical values}
}
\description{
"Round" a vector of numerical values to the closest element of a
set of discrete  values.
}
