% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_helpers.R
\name{side_by_side_table}
\alias{side_by_side_table}
\title{Layout helpers}
\usage{
side_by_side_table(...)
}
\arguments{
\item{...}{shiny widgets with a name, e,.g. \code{freq <- sliderInput('one', ....)}}
}
\description{
Layout helpers
}
