% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{smaller_select}
\alias{smaller_select}
\title{A few custom widgets}
\usage{
smaller_select(id, choices)
}
\arguments{
\item{id}{character string id to be assigned to the widget}

\item{choices}{list or vector of choices for the selectInput}
}
\description{
Create a smaller version of a shiny selectInput
}
