% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_calcs.R
\name{t_test_calcs}
\alias{t_test_calcs}
\title{Two-sample t-test plot}
\usage{
t_test_calcs(formula, data, level = 0.95, show_mean = TRUE,
  show_ci = TRUE, show_t = TRUE, var_equal = TRUE, y_range = NULL)
}
\description{
Two-sample t-test plot
}
