% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_explain.R
\name{ui_explain_tab}
\alias{ui_explain_tab}
\title{Create an explain tab}
\usage{
ui_explain_tab(title = "", icon = "bicycle")
}
\arguments{
\item{title}{A (short) character string naming the tab.}

\item{icon}{An optional icon, e.g. `"bicycle"`. Set to NULL to disable}
}
\description{
An explain tab containing an htmlOutput
}
