% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_main.R
\name{ui_main}
\alias{ui_main}
\title{Layout for the main app}
\usage{
ui_main(top_controls, ...)
}
\arguments{
\item{top_controls}{contents of the top block of controls}

\item{...}{tabs to  be inserted into  the app}
}
\description{
A Little App  consists of a top block of controls that stay visible aways, and
a set of tabs that implement different views: e.g. select data, compare current
to frozen plot, show statistics, show a description of the app, ...
}
