% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_top_controls.R
\name{ui_top_controls}
\alias{ui_top_controls}
\title{Create controls at top of app}
\usage{
ui_top_controls(title = "App controls", ...)
}
\arguments{
\item{title}{character string. The name to be used for the button that brings
up the modal containing the app-specific controls}

\item{...}{UI controls to be added on top.}
}
\description{
As a matter of design policy, each Little App has a row of controls
across the top of the window. These controls are always shown, irrespective
of which tab is showing. `ui_top_controls` is a function that creates the set of controls,
which then needs to be placed in an appropriate container such as a `miniUI::miniButtonBlock`.
}
