% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_compare_plot.R
\name{construct_compare_plot}
\alias{construct_compare_plot}
\title{Lay out the plots for the compare tab}
\usage{
construct_compare_plot(calc, what)
}
\arguments{
\item{calc}{the results  of the main (or frozen) calculation}

\item{what}{whether to display the "data plot" or the "model values"
side plot,  or both}
}
\description{
For internal use ... to lay out the main and side plots
for each of the current or frozen sides of the "compare" tab.
}
