% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_model_formula}
\alias{get_model_formula}
\title{construct the model formula}
\usage{
get_model_formula(sample, model_order = 1, interaction = TRUE,
  model_type = "lm")
}
\description{
construct the model formula
}
