% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_panels.R
\name{graph_panel}
\alias{graph_panel}
\title{panels for the UI}
\usage{
graph_panel()
}
\description{
Functions to make tabs for the UI: \code{graph_panel()},  \code{compare_panel()}, \code{stats_panel()},
\code{debug_panel()}
}
