% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_explain.R
\name{ui_explain_tab}
\alias{ui_explain_tab}
\title{Create an explain tab}
\usage{
ui_explain_tab(title = "", icon = "bicycle")
}
\arguments{
\item{icon}{An optional icon, e.g. `icon("bicycle")`. Set to NULL to disable}

\item{text}{A (short) character string naming the tab.}

\item{text}{A character string containing HTML.}
}
\description{
An explain tab shows a string that will be rendered as HTML
}
